ui/streamlit_app/requirements.txt
streamlit==1.24.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
pillow==9.5.0
openpyxl==3.1.2
xlrd==2.0.1
python-dotenv==1.0.0
pyjwt==2.8.0
streamlit-authenticator==0.2.2
watchdog==3.0.0

